added explicit  type annotations where possible

+++ Classes/DnDSession.py
@@ -5,5 +5,5 @@ class DnDSession:
   A DnD session class. Stores information about the current session.
   """
   is_active: bool = False
-  start_time : int = 0
+  start_time : float = 0
   current_session_dies : list[Dice] = []
